requirements.txt
cudatoolkit
python
pyyaml
tqdm
pytorch
numpy
pandas
scipy
